move IDPayload to types dir
