+++ backend/src/entities/posts/transaction.ts
@@ -1,8 +1,9 @@
 import { PostgresJsDatabase } from "drizzle-orm/postgres-js";
-import { CreatePostPayload, IDPayload, Media, PostWithMedia, UpdatePostPayload } from "./validator";
+import { CreatePostPayload, Media, PostWithMedia, UpdatePostPayload } from "./validator";
 import { groupsTable, mediaTable, membersTable, postsTable } from "../schema";
 import { eq, and, sql } from "drizzle-orm";
 import { ForbiddenError, NotFoundError } from "../../utilities/errors/app-error";
+import { IDPayload } from "../../types/id";
 
 export interface PostTransaction {
   createPost(post: CreatePostPayload): Promise<PostWithMedia | null>;
