+++ backend/src/entities/posts/validator.ts
@@ -6,6 +6,7 @@ import {
   TEXT_MAX_LIMIT,
 } from "../../constants/database";
 import { mediaTable, postsTable } from "../schema";
+import { IDPayload } from "../../types/id";
 
 export const createPostValidate = z
   .object({
@@ -36,10 +37,6 @@ export type CreatePostPayload = z.infer<typeof createPostValidate> & {
 };
 
 export type UpdatePostPayload = z.infer<typeof updatePostValidate> & IDPayload;
-export type IDPayload = {
-  userId: string;
-  id: string;
-};
 
 export type Media = typeof mediaTable.$inferSelect;
 type Post = typeof postsTable.$inferSelect;
