+++ backend/src/types/id.ts
@@ -1,0 +1,4 @@
+export type IDPayload = {
+  userId: string;
+  id: string;
+};
